➖ Disabling the preemption and pod priority by default
We currently are not using pod-priority and preemption. And according to the [changelog](https://github.com/kubernetes/kubernetes/blob/master/CHANGELOG-1.11.md#sig-scheduling) Pod Priority and Preemption has graduated to Beta, so it is enabled by default. Note that this involves significant and important changes for operators. The team also worked on improved performance and reliability of the scheduler.